+++ include/d3d12translationlayer/BlockAllocators.h
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 #pragma once
 
+#include "D3D12TranslationLayerDependencyIncludes.h"
 #include <memory>
 #include <set>
