--- include/d3d12translationlayer/D3D12TranslationLayerIncludes.h
@@ -1,32 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-#pragma once
-
-//Library Headers
-#include "SubresourceHelpers.hpp"
-#include "Util.hpp"
-#include "DeviceChild.hpp"
-
-#include <BlockAllocators.h>
-#include "Allocator.h"
-#include "XPlatHelpers.h"
-
-#include <formatdesc.hpp>
-
-#include "Sampler.hpp"
-#include "View.hpp"
-#include "PipelineState.hpp"
-#include "Fence.hpp"
-#include "Residency.h"
-#include "ResourceState.hpp"
-#include "RootSignature.hpp"
-#include "Resource.hpp"
-#include "Query.hpp"
-#include "ImmediateContext.hpp"
-#include "CommandListManager.hpp"
-
-#include "View.inl"
-#include "Sampler.inl"
-#include "ImmediateContext.inl"
-#include "CommandListManager.inl"
-#include <BlockAllocators.inl>
