+++ include/d3d12translationlayer/Fence.hpp
@@ -2,6 +2,10 @@
 // Licensed under the MIT License.
 #pragma once
 
+#include "D3D12TranslationLayerDependencyIncludes.h"
+#include "DeviceChild.hpp"
+#include "Util.hpp"
+
 namespace D3D12TranslationLayer
 {
     enum FENCE_FLAGS
