+++ include/d3d12translationlayer/RootSignature.hpp
@@ -2,6 +2,8 @@
 // Licensed under the MIT License.
 #pragma once
 
+#include "D3D12TranslationLayerDependencyIncludes.h"
+#include "DeviceChild.hpp"
 #include <array>
 
 namespace D3D12TranslationLayer
