+++ include/d3d12translationlayer/View.hpp
@@ -2,6 +2,9 @@
 // Licensed under the MIT License.
 #pragma once
 
+#include "D3D12TranslationLayerDependencyIncludes.h"
+#include "DeviceChild.hpp"
+#include "SubresourceHelpers.hpp"
 #include <bitset>
 
 namespace D3D12TranslationLayer
@@ -9,7 +12,6 @@ namespace D3D12TranslationLayer
     class Resource;
     //==================================================================================================================================
     // View
-    // Stores data responsible for remapping D3D11 views to underlying D3D12 views
     //==================================================================================================================================
 
     // These types are purely used to specialize the templated
@@ -20,12 +22,6 @@ namespace D3D12TranslationLayer
     template< class TIface >
     struct CViewMapper;
 
-    struct D3D12_UNORDERED_ACCESS_VIEW_DESC_WRAPPER
-    {
-        D3D12_UNORDERED_ACCESS_VIEW_DESC m_Desc12;
-        UINT m_D3D11UAVFlags;
-    };
-
 #define DECLARE_VIEW_MAPPER(View, DescType12, TranslationLayerDesc) \
     template<> struct CViewMapper<##View##Type> \
     { \
@@ -34,15 +30,8 @@ namespace D3D12TranslationLayer
     static decltype(&ID3D12Device::Create##View) GetCreate() { return &ID3D12Device::Create##View; } \
     }
 
-#define DECLARE_VIEW_MAPPER1(View, DescType, TranslationLayerDesc) \
-    template<> struct CViewMapper<##View##Type> \
-    { \
-    typedef TranslationLayerDesc TTranslationLayerDesc; \
-    typedef DescType TDesc12; \
-    }
-
     DECLARE_VIEW_MAPPER(ShaderResourceView, SHADER_RESOURCE_VIEW_DESC, D3D12_SHADER_RESOURCE_VIEW_DESC);
-    DECLARE_VIEW_MAPPER(UnorderedAccessView, UNORDERED_ACCESS_VIEW_DESC, D3D12_UNORDERED_ACCESS_VIEW_DESC_WRAPPER);
+    DECLARE_VIEW_MAPPER(UnorderedAccessView, UNORDERED_ACCESS_VIEW_DESC, D3D12_UNORDERED_ACCESS_VIEW_DESC);
 #undef DECLARE_VIEW_MAPPER
 
     class ViewBase : public DeviceChild
@@ -110,27 +99,6 @@ namespace D3D12TranslationLayer
     typedef View<ShaderResourceViewType> TSRV;
     typedef View<UnorderedAccessViewType> TUAV;
 
-    // Counter and Append UAVs have an additional resource allocated
-    // to hold the counter value
-    class UAV : public TUAV
-    {
-    public:
-        UAV(ImmediateContext* pDevice, const TTranslationLayerDesc &Desc, Resource &ViewResource) noexcept(false);
-        ~UAV() noexcept(false);
-
-        //Note: This is hiding the base class implementation not overriding it
-        void UsedInCommandList(UINT64 id)
-        {
-            TUAV::UsedInCommandList(id);
-            DeviceChild::UsedInCommandList(id);
-        }
-
-        static UAV *CreateView(ImmediateContext* pDevice, const TTranslationLayerDesc &Desc, Resource &ViewResource) noexcept(false) { return new UAV(pDevice, Desc, ViewResource); }
-
-    public:
-        UINT m_D3D11UAVFlags;
-    };
-
     class CDescriptorHeapManager;
     struct DescriptorHeapEntry
     {
@@ -144,4 +112,5 @@ namespace D3D12TranslationLayer
     };
 
     typedef TSRV SRV;
+    typedef TUAV UAV;
 };
