--- include/d3d12translationlayer/pch.h
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-#pragma once
-
-#include <D3D12TranslationLayerDependencyIncludes.h>
-#include <D3D12TranslationLayerIncludes.h>
