+++ src/d3d12translationlayer/PipelineState.cpp
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-#include "pch.h"
+#include "ImmediateContext.hpp"
+#include "RootSignature.hpp"
+#include "PipelineState.hpp"
 
 namespace D3D12TranslationLayer
 {
