+++ src/d3d12translationlayer/View.cpp
@@ -1,23 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#include "pch.h"
+#include "Resource.hpp"
+#include "View.hpp"
 
 namespace D3D12TranslationLayer
 {
-
     void ViewBase::UsedInCommandList(UINT64 id) 
     {
         if (m_pResource) { m_pResource->UsedInCommandList(id); }
-    }
-
-    //----------------------------------------------------------------------------------------------------------------------------------
-    UAV::UAV(ImmediateContext* pDevice, const TTranslationLayerDesc &Desc, Resource &ViewResource) noexcept(false)
-        : TUAV(pDevice, Desc.m_Desc12, ViewResource), m_D3D11UAVFlags(Desc.m_D3D11UAVFlags)
-    {
-    }
-
-    UAV::~UAV() noexcept(false)
-    {
     }
 };
